<div> Element with Class: Represents a container for the error page content. It has the class name "not-found-container".
Error Message: An <h1> heading element is used to display the error message "Sorry, the page you were looking for was not found."

Return to Home Link: A Link component is used to create a link to the home page ("/"). The link text is "Return to Home", and it has the class name "link-button".

Export Default: The NotFound component is exported as the default export of the module, making it available for use in other parts of the application