AuthRequired Function Component: The AuthRequired component is declared as a functional component using the arrow function syntax (() => { ... }).
Authentication Check: Inside the component, there's a check for authentication. In this case, the authenticated variable is set to false, indicating that the user is not authenticated.

Return JSX: If the user is not authenticated (authenticated is false), the component returns a Navigate component. This causes a redirect to the login page ("/login") with state data containing a message indicating that the user must log in first.

Return Outlet: If the user is authenticated, the component returns an Outlet. The Outlet component is used to render child routes, allowing access to the protected content